### 2025-08-15 #8799 @Kevin-Mattheus-Moerman
@yjmantilla thanks for this submission. As you can see I've just flagged this for scope. This is because I need some help from the editorial board to see if the functionality offered here is sufficient and in scope for JOSS. We should get back to you about this within about 2 weeks.
