### 2025-09-15 #8799 @Kevin-Mattheus-Moerman
@yjmantilla thanks for your patience. The board has determined this work may be in scope for JOSS. I'll remove the query-scope label and we'll proceed to finding an editor to handle this work. However, we are currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.  As such, for now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!
