### 2026-02-12 7a1fddd
Add Research Impact Statement and AI Usage Disclosure sections to paper.md
