### 2026-02-19 91232db
Specify minimum version for fastapi-jsonrpc in requirements files
