### 2026-02-19 74248b0
Update Python version in CI workflow to 3.11
