### 2026-02-19 643ada6
Specify datatype and suffix for BIDSPath in apply_rules_to_single_file function
