### 2026-02-19 0ec2c8f
Ensure data directory is created in test_sova2coin function
