### 2026-02-19 3670aae
Convert raw filename to string if it's a Path object in apply_rules_to_single_file function
