### 2026-02-19 16bba7f
Update import for MutableMapping to use collections.abc in flatten function
