### 2026-02-20 #8799 @rkurchin
Thanks, @yjmantilla, as you can see above, we do still need that "State of the field" section as well – see more details in the docs [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain).
