### 2026-02-26 64b0461
Update publication date and enhance summary in paper.md
