+++ paper/paper.md
@@ -55,13 +55,13 @@ affiliations:
     index: 9
 
 
-date: 05 August 2025
+date: 25 February 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-Electroencephalography (EEG) data are used in many fields, from neuroscience to clinical research, but it often comes in different formats and structures, making it hard to organize, share, or compare across studies. SOVABIDS is an open-source tool that helps researchers convert EEG data into the Brain Imaging Data Structure (BIDS) [@bids], a standard data structure that aligns with FAIR principles (Findability, Accessibility, Interoperability, and reusability). Specifically, BIDS ensures data consistency and interoperability, making it easier to analyze, share, and integrate EEG data across different tools, data repositories and research groups. Rather than manually renaming files or reorganizing folders, SOVABIDS allows users to define simple rules to automate the conversion (\autoref{fig:use} illustrates this conversion process). The tool is designed to be flexible and user-friendly, supporting customization without requiring advanced programming skills. It can be used as a Python package or as a command-line tool, and includes [comprehensive documentation](https://sovabids.readthedocs.io/en/latest/README.html) with [tutorials](https://sovabids.readthedocs.io/en/latest/auto_examples/index.html) to help users get started. SOVABIDS also integrates with other tools, making it suitable for both small research projects and large collaborative studies.
+Electroencephalography (EEG) data are used in many fields, from neuroscience to clinical research, but recordings are stored in a wide variety of formats and lab-specific folder structures, making them difficult to organize, share, or compare across studies. SOVABIDS is an open-source tool that helps researchers convert EEG data into the Brain Imaging Data Structure (BIDS) [@bids], a standard that promotes FAIR data practices (Findability, Accessibility, Interoperability, and Reusability) and is increasingly expected for data sharing and large-scale collaboration. Rather than manually renaming files or writing dataset-specific scripts, SOVABIDS allows users to define conversion rules in human-readable YAML configuration files, which are then applied automatically across all participants in a dataset (\autoref{fig:use} illustrates this process). The tool is usable as a Python package or command-line tool, and its API supports integration with graphical frontends for users who prefer a visual interface. Comprehensive documentation and tutorials are available at [sovabids.readthedocs.io](https://sovabids.readthedocs.io/en/latest/README.html).
 
 ![Illustration of the EEG to BIDS conversion. The left side shows raw EEG files with participant-specific naming conventions (for example, P1_S0_EC.cnt), where P1 and P2 represent participants, S0 and S1 indicate sessions, and EC (Eyes Closed) and EO (Eyes Open) refer to tasks. These raw files are converted into the BIDS format, shown on the right, where data are systematically organized into subject (sub-), session (ses-), and modality (eeg) folders. Each EEG recording is saved in standardized BIDS-compliant formats, including .edf for EEG signals and .tsv/.json for metadata. \label{fig:use}](main-use.png)
 
@@ -69,66 +69,75 @@ Electroencephalography (EEG) data are used in many fields, from neuroscience to 
 
 Electroencephalography is a widely used neuroimaging technique that provides high temporal resolution for studying brain activity. Its applications span numerous fields, including cognitive neuroscience, clinical diagnostics, brain-computer interfaces, and neuroengineering. With the increasing volume and complexity of EEG data, ensuring reproducibility, standardization, and interoperability has become a growing priority in the field. The Brain Imaging Data Structure for EEG (EEG-BIDS) [@eegbids] provides a consistent framework for organizing EEG datasets, facilitating data sharing [@openneuro], large-scale collaborations, cross-study comparisons, and promoting FAIR data practices [@fairdata] across a wide range of research applications from fundamental cognitive neuroscience to large-scale clinical neuroimaging.
 
[…]
