### 2026-02-26 6c7ea66
Clarify the use of generative AI tools in project maintenance and support tasks in paper.md
