### 2026-02-26 e8e6b74
Fix punctuation in generative AI tools usage description in paper.md
