### 2026-02-26 bc04249
Refine description of AI tools used for maintenance and support tasks in paper.md
