### 2026-03-05 #8799 @qian-chu
> 👋 [@tmedani](https://github.com/tmedani), [@qian-chu](https://github.com/qian-chu), and/or [@wmvanvliet](https://github.com/wmvanvliet), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi, thanks for contacting me! I'm more than happy to review this submission.
