### 2026-03-06 #10146 @editorialbot
[REVIEW]: SOVABIDS: EEG-to-BIDS conversion software focused on automation, reproducibility and interoperability
**Submitting author:** @yjmantilla (<a href="http://orcid.org/0000-0003-4473-0876">Yorguin Mantilla</a>)
**Repository:** https://github.com/yjmantilla/sovabids
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.2
