### 2026-03-06 yjmantilla/sovabids#74 @qian-chu
Dependency management
Part of review at https://github.com/openjournals/joss-reviews/issues/10146

Thanks for the great work on the package! I’ve begun testing and encountered my first error, which appears to be related to a missing dependency:

```
>>> sovapply source_path bids_path rules_path
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "...\AppData\Local\miniforge3\envs\sovabids\Scripts\sovapply.exe\__main__.py", line 2, in <module>
    from sovabids.rules import sovapply
  File "...\AppData\Local\miniforge3\envs\sovabids\Lib\site-packages\sovabids\rules.py", line 16, in <module>
    from pandas import read_csv
ModuleNotFoundError: No module named 'pandas'
```

Seems pandas was not listed in requirements-dev.txt, requirements-user.txt, or requirements-gui.txt
