### 2026-03-07 yjmantilla/sovabids#74 @qian-chu
In addition, setup.cfg requries python>=3.8. However, fastapi-jsonrpc>=3.4.2 (from requirements-gui.txt) requires at least python=3.9.

This highlights somewhat incomplete dependency management. Maybe you can consider using some tools and test installation on multiple python versions and OS systems with CI matrix and GitHub workflow?
