### 2026-03-09 yjmantilla/sovabids#75 @wmvanvliet
[joss-review] Some recommendations regarding the installation process.
sovabids would be much easier to install if it could offer a pip package. Also, the old `setup.py`/`setup.cfg` system is being phased out. I recommend moving to the `pyproject.toml` system (see https://packaging.python.org for a detailed tutorial on how this works). This way, you could also describe the dependencies inside `pyproject.toml` and the installation instructions would be more in line with other python projects, namely:

For users: `pip install sovabids`

For developers:
```bash
git clone https://github.com/yjmantilla/sovabids
cd sovabids
pip install -e .`
```

at the moment, you've included `-e .` in `requirements-gui.txt` which is not normally done and breaks the contract of a `requirements.txt` file a little, in that it only describes dependencies.
