### 2026-03-09 eab461b
Output log information to stdout
