+++ sovabids/convert.py
@@ -80,12 +80,20 @@ def convert_them(mappings_input):
 def sovaconvert():
     """Console script usage for conversion."""
     # see https://github.com/Donders-Institute/bidscoin/blob/master/bidscoin/bidsmapper.py for example of how to make this
+    logging.basicConfig(format="%(message)s")
+    LOGGER.setLevel(logging.WARN)
+
     parser = argparse.ArgumentParser()
     subparsers = parser.add_subparsers()
 
     parser = subparsers.add_parser('convert_them')
     parser.add_argument('mappings',help='The mapping file of the conversion.')
+    parser.add_argument('-v','--verbose', action="store_true", help='Make the output more verbose.')
     args = parser.parse_args()
+
+    if args.verbose:
+        LOGGER.setLevel(logging.INFO)
+
     convert_them(args.mappings)
 
 if __name__ == "__main__":
