### 2026-03-10 #8799 @tmedani
> 👋 [@tmedani](https://github.com/tmedani), [@qian-chu](https://github.com/qian-chu), and/or [@wmvanvliet](https://github.com/wmvanvliet), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @rkurchin
I missed the notifications earlier... my apologies.
Please let me know if you still need an additional reviewer.
