### 2026-03-16 yjmantilla/sovabids#77 @qian-chu
Support for `Path` objects in file path arguments
Part of review at https://github.com/openjournals/joss-reviews/issues/10146

I noticed that the package currently appears to accept file paths only as strings. Would it be possible to also support `pathlib.Path` objects? These are preferred by some users and are commonly used in navigating data folders.
