### 2026-03-16 #10146 @qian-chu
# Comments

I was very pleased to review `sovabids`. In particular, I commend the authors’ efforts to design the software around extracting, logging, and applying conversion entities and rules in a transparent and reproducible way. Source to raw data conversion is often handled behind the scenes in BIDS workflows, and `sovabids` represents a meaningful step forward in making these processes explicit. This benefits the users, who can track what kind of manual adjustments they made (e.g., channel renaming), but also downstream users of the dataset, who gain clearer insight into how the raw data were processed and converted to BIDS format.

Moreover, `path_analysis` is a novel and powerful approach for extracting BIDS entities from file paths. It provides a well-structured implementation of practices that are common in BIDS conversion but are often coded by researchers from scratch.

I have completed the checklist above. There are only a few boxes I haven't ticked, reflecting areas (detailed below) where I believe the authors could make further improvements. I would of course be happy to check these items if the authors address the points or satisfactorily clarify them in their response.

## Conceptual framing of `sovabids`

One aspect that I think could be strengthened is the articulation of why `sovabids` stands out from existing tools. In `Statement of the Field`, the paper notes that several other packages “require dataset‑specific scripting,” but it is not entirely clear how `sovabids` avoids this requirement.

- With respect to versatility, the YAML‑based rules and mappings still appear to be dataset‑specific, as each dataset typically has its own file structure. If the authors are referring to the `path_analysis` component, while it does provide flexible mechanisms for extracting BIDS entities from file paths (via regular expressions, placeholders, or paired examples), it nonetheless requires dataset‑specific patterns. From this perspective, it may be more accurate to frame `sovabids` as providing a flexible, semi‑automatic API for inferring subject‑ and run‑level information, rather than eliminating dataset‑specific configuration altogether. The key distinction, as I see it, is that users no longer need to manually extract and pass this information when converting each subject and run.

- With respect to scripting, while YAML files can be edited manually (including via the `sovabids` GUI), this approach introduces the need to learn a file format that is not otherwise exposed to BIDS users. In practice, users are also likely to generate rule files programmatically (e.g., from Python dictionaries, as shown in the [tutorial](https://sovabids.readthedocs.io/en/latest/auto_examples/example_heuristic_source_target_pair.html)). My personal impression is that this may actually introduce a slightly higher initial learning barrier, albeit one that comes with clear benefits.

Therefore, I recommend the authors to reconsider framing `sovabids`' advantages in `State of the field` and `Software Design`. This is not to suggest that sovabids does not compete with other tools, nor that its use of rules and mappings is misguided. Rather, a more specific and nuanced framing could better reflect the practical differences between sovabids and other tools, and more clearly articulate the concrete advantages it offers to users.

## Installation

The installation instructions are generally clear. However, I encountered difficulties when attempting to install the full set of dependencies using the provided requirements.txt file (see https://github.com/yjmantilla/sovabids/issues/74).

More broadly, I found the use of multiple requirements files somewhat confusing (e.g., difference between requirements.txt and requirements-user.txt). I would recommend that the authors consider consolidating dependencies into a single requirements specification with optional dependencies (e.g., via `extras_require` in setup.py or the equivalent in `pyproject.toml`, as recommended by @wmvanvliet at https://github.com/yjmantilla/sovabids/issues/75). Adopting this approach would simplify dependency management and align the project with current practices in Python packaging.

Subsequently, this would also streamline installation via pip. For example, the installation process could be reduced to a single command such as:

```bash
pip install sovabids[gui]
```

if the package will be released on PyPI, or:

```bash
pip install "sovabids[gui] @ https://github.com/yjmantilla/sovabids.git"
```

to allow users to install the latest version directly from GitHub.

I would also recommend explicitly specifying supported versions of MNE and MNE‑BIDS. In my personal opinion, MNE ≥ 1.8 would be a sensible lower bound, as several export‑related issues present in earlier versions have since been resolved. I personally contributed fixes for some of these issues, for example:

- https://github.com/mne-tools/mne-python/pull/12656
- https://github.com/mne-tools/mne-python/pull/12676

## Implementation

I tested the script and GUI and they functioned as expected with the example dataset. I have a few suggestions, which the authors may consider:

1. Currently, if the `path_analysis` finds no matching files to convert, it doesn't log or print any error, instead leaving `Individual: []` in `mapping.yaml`. I find this behavior unhelpful, as users may not realize that no files were matched. It would be preferable to explicitly inform users when no files are found for conversion.
2. Currently, errors are written to `sovabids.log`. First, they should probably be in `sovabids.log.errors` instead. Second, I recommend also printing errors to the console to make it easier for users to identify and debug issues during the conversion process. Alternatively, users should be more explicitly informed about the existence and location of the log file, and guided on how to consult it when errors occur.
3. The GUI is great for selecting files to convert, manually tweaking the mappings, and generating dataset for download. The only aspect that feels slightly unintuitive is that users must either load an existing rules file or create one from scratch. A useful enhancement would be for the GUI to provide modular components to help users generate rules incrementally, or to offer an empty (or minimally populated) template that users can fill in.
4. As raised in https://github.com/yjmantilla/sovabids/issues/77: it would be nice to also accept `Path` objects.

## Documentation

The documentation is informative and generally clear. I have a few minor suggestions for improvement:

First, the overall ordering and inter‑linking of the documentation could be improved. When learning how to use the package, I found myself jumping back and forth between sections to piece together the workflow. In practice, I followed this sequence:

1. [Intro to sovabids](https://sovabids.readthedocs.io/en/latest/README.html)
2. [Rules File Schema](https://sovabids.readthedocs.io/en/latest/rules_schema.html) (after which I prepared a YAML file)
3. [Intro to sovabids](https://sovabids.readthedocs.io/en/latest/README.html) (to understand how to use `sovapply`)
4. [Mappings File Schema](https://sovabids.readthedocs.io/en/latest/mappings_schema.html) (to compare my mapping file with the documentation)
5. [Intro to sovabids](https://sovabids.readthedocs.io/en/latest/README.html) (to see how to use `sovaconvert`)

This could be improved by providing a more linear workflow, presumably in the form of a quickstart section where the LEMON example is shown. Along the way, you can provide link to the rules and mapping files as well as the API (e.g., `sovapply` and `sovaconvert`).

In [Rules File Schema](https://sovabids.readthedocs.io/en/latest/rules_schema.html#), it might be useful to list the file names that will be generated. Terms like `sidecar` might still be a bit technical for users who are new to BIDS compared to the more intuitive destination file (e.g., `*_eeg.json`).

I also echo with this issue: https://github.com/yjmantilla/sovabids/issues/60. Although it was mentioned in the paper that BrainVision and CNT are the one tested, I think readers will curious to know what other EEG formats can be used by sovabids.

Also a tiny neatpick: the authors mentioned multiple times that `mappings.yaml` is "derived" from `rules.yaml`. I personally prefer "populated" as a more intuitive term.

## References

To illustrate that `sovabids` is an officially supported tool in `neurodesk`, authors may consider citing an additional paper where `sovabids` was mentioned: https://doi.org/10.52294/001c.144107.
