### 2026-04-30 3885851
Implement unicode dash handling in CLI arguments and add tests
