### 2026-04-30 e7a7a80
Enhance path handling by accepting pathlib.Path in various functions and updating documentation
