### 2026-04-30 65416c6
Update Python version to 3.11 in Read the Docs configuration
