+++ .readthedocs.yml
@@ -14,7 +14,7 @@ sphinx:
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.8"
+    python: "3.11"
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
