### 2026-04-30 yjmantilla/sovabids#77 @yjmantilla
I added pathlib.Path support in #79 (already merged)
