+++ requirements-user.txt
@@ -1,5 +1,5 @@
-bids_validator
-pybv >= 0.7.3
+bids-validator
+pybv>=0.7.3
 mne-bids
 pyyaml
-.
+pandas
