+++ requirements-gui.txt
@@ -1,11 +1,11 @@
-bids_validator
+bids-validator
 flask
 uvicorn
 fastapi
-fastapi-jsonrpc >= 3.4.2
-pybv >= 0.7.3
+fastapi-jsonrpc>=3.4.2
+pybv>=0.7.3
 httpx
 mne-bids
 pyyaml
+pandas
 git+https://github.com/yjmantilla/bidscoin.git@sovabids
-.
