+++ .github/workflows/python-tests.yml
@@ -1,6 +1,3 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
 name: python-tests
 
 on:
@@ -16,37 +13,53 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [3.11]
-        node:  [18] # might need to update this manually?
+        python-version: ["3.11"]
+        node: ["18"]
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v4
+      with:
+        fetch-depth: 0
+
+    - name: Install uv
+      uses: astral-sh/setup-uv@v5
+      with:
+        enable-cache: true
+
     - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
+
     - name: Use Node.js ${{ matrix.node }}
-      uses: actions/setup-node@v1
+      uses: actions/setup-node@v4
       with:
         node-version: ${{ matrix.node }}
+
     - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install pytest
-        if [ -f requirements-dev.txt ]; then pip install -r requirements-dev.txt; fi
-#        pip install -e .,already inside requirements-dev.txt
+      run: uv sync --frozen --all-extras
+
     - name: Test with pytest
-      run: |
-        pytest tests/ --cov=sovabids/ --cov-report=xml --verbose -s
+      run: uv run pytest tests/ --cov=sovabids/ --cov-report=xml --verbose -s
+
     - name: Build docs
       run: |
         cd docs
-        make clean
-        make html
+        uv run make clean
+        uv run make html
+
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v5
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         slug: yjmantilla/sovabids
-    - name : Test with web-validator
-      run: |
-        /bin/bash tests/test_web_validator.sh
+
+    - name: Upload uv.lock as artifact
+      if: success()
+      uses: actions/upload-artifact@v4
+      with:
+        name: uv-lock-${{ github.sha }}
+        path: uv.lock
+        retention-days: 90
+
+    - name: Test with web-validator
+      run: /bin/bash tests/test_web_validator.sh
