### 2026-04-30 a68cf33
Update installation requirements and contributing guidelines
