+++ requirements-user.txt
@@ -1,5 +1,0 @@
-bids-validator
-pybv>=0.7.3
-mne-bids
-pyyaml
-pandas
