### 2026-04-30 93b44b4
Update Python version matrix in CI workflow to include 3.9 and 3.10
Co-authored-by: Copilot <copilot@github.com>
