+++ .github/workflows/python-tests.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.11"]
+        python-version: ["3.9", "3.10", "3.11"]
         node: ["18"]
     steps:
     - uses: actions/checkout@v4
