### 2026-04-30 5551859
Add 'gui' to extra_requirements in Read the Docs configuration (bug in build)
