+++ .readthedocs.yml
@@ -18,3 +18,4 @@ python:
       path: .
       extra_requirements:
         - docs
+        - gui
