### 2026-04-30 496bec2
Refactor pyproject.toml to use hatchling for build system and update project metadata
