### 2026-05-01 1ea211a
fix missing bidscoin in docs and tests
