+++ .readthedocs.yml
@@ -11,6 +11,9 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.11"
+  jobs:
+    post_install:
+      - python -m pip install "git+https://github.com/yjmantilla/bidscoin.git@sovabids"
 
 python:
   install:
