+++ paper/paper.md
@@ -69,7 +69,7 @@ Electroencephalography (EEG) data are used in many fields, from neuroscience to 
 
 Electroencephalography is a widely used neuroimaging technique that provides high temporal resolution for studying brain activity. Its applications span numerous fields, including cognitive neuroscience, clinical diagnostics, brain-computer interfaces, and neuroengineering. With the increasing volume and complexity of EEG data, ensuring reproducibility, standardization, and interoperability has become a growing priority in the field. The Brain Imaging Data Structure for EEG (EEG-BIDS) [@eegbids] provides a consistent framework for organizing EEG datasets, facilitating data sharing [@openneuro], large-scale collaborations, cross-study comparisons, and promoting FAIR data practices [@fairdata] across a wide range of research applications from fundamental cognitive neuroscience to large-scale clinical neuroimaging.
 
-Despite these advantages, converting EEG datasets to BIDS in practice remains a significant challenge. EEG data varies widely in format due to the multitude of proprietary standards used by different hardware vendors, and datasets are often organized according to lab-specific or acquisition-driven conventions rather than any consistent structure. For researchers without strong programming backgrounds, particularly those at smaller or less well-resourced institutions, this conversion process typically requires either substantial manual effort or the ability to write dataset-specific scripts, both of which are error-prone and difficult to reproduce or scale to large multi-participant studies. SOVABIDS addresses this gap by providing a rule-based, semi-automated conversion workflow that is accessible to non-programmers while remaining flexible enough to handle the heterogeneity common in real-world EEG datasets.
+Despite these advantages, converting EEG datasets to BIDS in practice remains a significant challenge. EEG data varies widely in format due to the multitude of proprietary standards used by different hardware vendors, and datasets are often organized according to lab-specific or acquisition-driven conventions rather than any consistent structure. For researchers without strong programming backgrounds, particularly those at smaller or less well-resourced institutions, this conversion process typically requires either substantial manual effort or the ability to write dataset-specific scripts, both of which are error-prone and difficult to reproduce or scale to large multi-participant studies. SOVABIDS addresses this gap by providing a rule-based, semi-automated conversion workflow that is accessible to users with limited programming experience while remaining flexible enough to handle the heterogeneity common in real-world EEG datasets.
 
 # State of the Field
 
