### 2026-05-01 e8403c9
refine SOVABIDS approach description in paper.md for clarity and accessibility
