+++ paper/paper.md
@@ -85,7 +85,7 @@ Bidsme [@bidsme] is a general-purpose converter that assumes datasets are alread
 
 A natural question is whether SOVABIDS' goals could have been achieved by contributing to an existing tool, particularly MNE-BIDS. We argue they could not. MNE-BIDS is designed around a scripting paradigm where conversion logic is expressed in Python code. Adding rule-based, configuration-driven automation as a non-breaking extension would require architectural changes that diverge from MNE-BIDS’ current scripting-oriented design philosophy and typical usage patterns. The same barrier applies to FieldTrip and EEGLAB, where conversion is tightly coupled to their respective analysis environments.
 
-SOVABIDS instead introduces a distinct approach: an explicit two-tier separation between dataset-level rules and participant-level mappings, encoded in human-readable YAML configuration files. This separation enables automated propagation of general conversion logic while systematically capturing participant-specific deviations without requiring any scripting. Combined with flexible metadata extraction from arbitrary file paths, this design allows SOVABIDS to handle the heterogeneous directory structures and vendor-specific naming conventions common in real-world EEG datasets, while remaining accessible to researchers without programming experience. It also preserves full provenance, enabling reproducible and auditable conversion workflows.
+SOVABIDS instead introduces a distinct approach: an explicit two-tier separation between dataset-level rules and participant-level mappings, encoded in human-readable YAML configuration files. Like other tools, SOVABIDS requires dataset-specific configuration, but this configuration is defined once at the dataset level and then automatically propagated to generate participant-level mappings, eliminating the need for per-file scripting or manual intervention. Combined with a flexible, semi-automatic API for extracting BIDS entities from arbitrary file paths (via paired examples, placeholder templates, or regular expressions), this design allows SOVABIDS to handle the heterogeneous directory structures and vendor-specific naming conventions common in real-world EEG datasets. While the YAML-based approach introduces its own modest learning requirement, it offers a declarative alternative to general-purpose scripting that is inspectable, shareable, and auditable, preserving full provenance for reproducible conversion workflows.
 
 # Software Design
 
