### 2026-05-01 b7cdd0f
enhance accessibility section in paper.md by clarifying YAML configuration benefits for non-technical users
