### 2026-05-01 1d632c8
refine metadata extraction description in paper.md for clarity and accuracy
