+++ paper/paper.md
@@ -106,7 +106,7 @@ EEG experiments typically produce multiple identically-organised datasets, one p
 
 This two-tier approach, inspired by tools such as Bidscoin [@bidscoin] and HeuDiConv [@heudi] (both focused on MRI), extends their model by explicitly supporting non-identical participant structures. Users can also derive an initial Rules File from a community or institutional template to further reduce manual input, and can connect an external GUI via SOVABIDS' API for supervised adjustment of edge cases where full automation is not possible.
 
-To support flexible metadata extraction, SOVABIDS implements heuristics based on common EEG file path patterns. Rather than requiring data to be pre-organized into a standard folder hierarchy before conversion can begin, as both Bidscoin and Bidsme do, SOVABIDS can extract subject, session, task, and other BIDS-relevant properties from arbitrary file paths through flexible pattern matching. This is supported through three approaches of increasing technicality: paired source-target examples for users unfamiliar with pattern matching, placeholder-based templates for intermediate users, and full regular expressions for advanced users who need precise control.
+To support flexible metadata extraction, SOVABIDS provides a semi-automatic API for inferring subject, session, task, and other BIDS-relevant properties from arbitrary file paths. Rather than requiring data to be pre-organized into a standard folder hierarchy before conversion can begin, as both Bidscoin and Bidsme do, SOVABIDS extracts these properties through pattern matching that is defined once at the dataset level and applied automatically across all files. This is supported through three approaches of increasing technicality: paired source-target examples for users unfamiliar with pattern matching, placeholder-based templates for intermediate users, and full regular expressions for advanced users who need precise control.
 
 ## 3. Reproducible conversion
 
