### 2026-05-01 yjmantilla/sovabids#75 @yjmantilla
Thanks for pointing this out, I have added sovabids as a PyPI package (https://pypi.org/project/sovabids/ ). It is now easily installable and uses pyproject.toml to define the different dependency requirements.

Relevant PRs #80 and #81.
