### 2026-05-01 yjmantilla/sovabids#74 @yjmantilla
Thanks for the suggestions!

I have added pandas to the dependencies ( now on pyproject.toml !) and test on 3.9, 3.10, and 3.11 using a CI Matrix.
