### 2026-05-04 20ec4c8
update dependencies in pyproject.toml to specify minimum versions for mne-bids and mne
