### 2026-05-04 0ce3dbc
add warning for no matching files in apply_rules and enhance tests for logging
