### 2026-05-04 be3281b
refactor logging setup and add comprehensive tests for stderr handling and error file creation
