### 2026-05-04 4c70039
Add Terminal User Interface (TUI) for EEG to BIDS conversion
