+++ sovabids/tui.py
@@ -1,0 +1,509 @@
+"""Terminal User Interface for sovabids."""
+from __future__ import annotations
+
+import os
+import threading
+from copy import deepcopy
+from typing import ClassVar
+
+import yaml
+from textual import work
+from textual.app import App, ComposeResult
+from textual.containers import Horizontal, ScrollableContainer, Vertical
+from textual.reactive import reactive
+from textual.screen import ModalScreen
+from textual.widgets import (
+    Button,
+    DataTable,
+    DirectoryTree,
+    Footer,
+    Header,
+    Input,
+    Label,
+    RichLog,
+    Rule,
+    Select,
+    Static,
+    TabbedContent,
+    TabPane,
+)
+
+from sovabids.settings import SUPPORTED_EXTENSIONS
+
+
+# ── Directory picker modal ────────────────────────────────────────────────────
+
+class DirPickerScreen(ModalScreen[str]):
+    """Modal that lets user browse and confirm a directory."""
+
+    DEFAULT_CSS = """
+    DirPickerScreen {
+        align: center middle;
+    }
+    DirPickerScreen > Vertical {
+        width: 80%;
+        height: 80%;
+        border: thick $primary;
+        background: $surface;
+        padding: 1;
+    }
+    DirPickerScreen DirectoryTree {
+        height: 1fr;
+    }
+    DirPickerScreen Label {
+        height: 1;
+        margin-bottom: 1;
+    }
+    DirPickerScreen Horizontal {
+        height: 3;
+        align: right middle;
+    }
+    DirPickerScreen Button {
+        margin-left: 1;
+    }
+    """
+
+    def __init__(self, start: str = ".") -> None:
+        super().__init__()
+        self._start = os.path.abspath(start) if start else os.path.expanduser("~")
+
+    def compose(self) -> ComposeResult:
+        with Vertical():
+            yield Label("Select directory (press Enter or OK to confirm):")
+            yield DirectoryTree(self._start, id="dir-tree")
+            with Horizontal():
+                yield Button("Cancel", id="cancel", variant="default")
+                yield Button("OK", id="ok", variant="primary")
+
+    def on_directory_tree_directory_selected(
+        self, event: DirectoryTree.DirectorySelected
+    ) -> None:
+        self._selected = str(event.path)
+        self.query_one("#ok", Button).label = f"OK  [{os.path.basename(str(event.path))}]"
+
+    def on_button_pressed(self, event: Button.Pressed) -> None:
+        if event.button.id == "cancel":
+            self.dismiss("")
+        elif event.button.id == "ok":
+            path = getattr(self, "_selected", "")
+            self.dismiss(path)
+
+
+# ── Setup tab ─────────────────────────────────────────────────────────────────
+
+class SetupPane(Static):
+    DEFAULT_CSS = """
+    SetupPane { padding: 1 2; }
+    SetupPane Label { margin-top: 1; }
+    SetupPane .path-row { height: 3; }
+    SetupPane Input { width: 1fr; }
+    SetupPane Button { width: 14; margin-left: 1; }
+    SetupPane .hint { color: $text-muted; margin-bottom: 1; }
+    """
+
+    def compose(self) -> ComposeResult:
+        yield Label("[b]Source directory[/b] — raw EEG files")
+        with Horizontal(classes="path-row"):
+            yield Input(placeholder="/path/to/source", id="source-input")
+            yield Button("Browse…", id="browse-source")
+        yield Label("[b]BIDS output directory[/b]")
+        with Horizontal(classes="path-row"):
+            yield Input(placeholder="/path/to/bids_output", id="bids-input")
+            yield Button("Browse…", id="browse-bids")
+        yield Rule()
+        yield Label("[b]Load existing rules file[/b] (optional — skips Rules tab)")
+        with Horizontal(classes="path-row"):
[…]
