+++ pyproject.toml
@@ -43,6 +43,7 @@ Source = "https://github.com/yjmantilla/sovabids"
 [project.scripts]
 sovapply = "sovabids.rules:sovapply"
 sovaconvert = "sovabids.convert:sovaconvert"
+sovatui = "sovabids.tui:main"
 
 [project.optional-dependencies]
 gui = [
@@ -52,6 +53,9 @@ gui = [
     "fastapi-jsonrpc>=3.4.2",
     "httpx",
     "pydantic",
+]
+tui = [
+    "textual>=0.50.0",
 ]
 docs = [
     "sphinx>=4.0.0",
@@ -66,7 +70,7 @@ docs = [
 dev = [
     "pytest",
     "pytest-cov",
-    "sovabids[gui,docs]",
+    "sovabids[gui,tui,docs]",
 ]
 
 [tool.hatch.metadata]
