### 2026-05-04 fdbbfb3
Refactor preview timer cancellation method and update package dependencies in uv.lock
