### 2026-05-04 9703145
Add FilesListScreen modal to display matched files and enhance RulesPane functionality
