+++ sovabids/tui.py
@@ -140,6 +140,47 @@ class SetupPane(Static):
         }
 
 
+# ── Files list modal ─────────────────────────────────────────────────────────
+
+class FilesListScreen(ModalScreen):
+    """Modal showing all matched files with full paths."""
+
+    DEFAULT_CSS = """
+    FilesListScreen { align: center middle; }
+    FilesListScreen > Vertical {
+        width: 92%; height: 90%;
+        border: thick $primary; background: $surface; padding: 1;
+    }
+    FilesListScreen DataTable { height: 1fr; }
+    FilesListScreen Horizontal { height: 3; align: right middle; }
+    """
+
+    def __init__(self, files: list[str]) -> None:
+        super().__init__()
+        self._files = files
+
+    def compose(self) -> ComposeResult:
+        with Vertical():
+            yield Label(f"[b]{len(self._files)} matched file(s)[/b] — full paths")
+            yield DataTable(id="files-table", zebra_stripes=True)
+            with Horizontal():
+                yield Button("Close", id="close-files", variant="primary")
+
+    def on_mount(self) -> None:
+        table = self.query_one("#files-table", DataTable)
+        table.add_column("File path")
+        for f in self._files:
+            table.add_row(f)
+
+    def on_button_pressed(self, event: Button.Pressed) -> None:
+        if event.button.id == "close-files":
+            self.dismiss()
+
+    def on_key(self, event) -> None:  # noqa: ANN001
+        if event.key == "escape":
+            self.dismiss()
+
+
 # ── Rules tab ─────────────────────────────────────────────────────────────────
 
 class RulesPane(Static):
@@ -147,21 +188,33 @@ class RulesPane(Static):
     RulesPane { padding: 1 2; }
     RulesPane Label { margin-top: 1; }
     RulesPane .hint { color: $text-muted; }
+    RulesPane #pattern-examples {
+        color: $text-muted;
+        margin-bottom: 1;
+        padding: 0 2;
+        border-left: solid $primary;
+    }
     RulesPane #preview-label {
         margin-top: 1;
         padding: 1;
         border: solid $primary;
         color: $success;
+        width: 1fr;
     }
     RulesPane #preview-label.error { color: $error; }
     RulesPane #preview-label.muted { color: $text-muted; }
+    RulesPane #show-files { width: 18; margin-left: 1; }
     RulesPane Select { margin-bottom: 1; }
     RulesPane Input { margin-bottom: 1; }
     RulesPane .section-head { margin-top: 1; }
     """
 
-    _preview_timer = None
     _preview_lock: ClassVar = threading.Lock()
+
+    def __init__(self) -> None:
+        super().__init__()
+        self._preview_timer = None
+        self._matched_files: list[str] = []
 
     def compose(self) -> ComposeResult:
         yield Label("[b]EEG file extension[/b]")
@@ -170,8 +223,20 @@ class RulesPane(Static):
 
         yield Label("[b]Path pattern[/b]")
         yield Static(
-            "Use %entity% placeholders, e.g.  …/%subject%_%task%_%run%.vhdr",
+            "Use %entity% placeholders matched against the full file path.",
             classes="hint",
+        )
+        yield Static(
+            "[dim]Examples:[/dim]\n"
+            "  [cyan]%subject%_%task%.vhdr[/cyan]\n"
+            "    matches: /data/001_rest.vhdr  →  subject=001, task=rest\n"
+            "  [cyan]%subject%/%task%_%run%.bdf[/cyan]\n"
+            "    matches: /data/001/rest_01.bdf  →  subject=001, task=rest, run=01\n"
[…]
