### 2026-05-04 4537f02
Enhance FilesListScreen and add ChannelNamesScreen modal for improved file extraction details and channel information display
