+++ sovabids/tui.py
@@ -20,6 +20,8 @@ from textual.widgets import (
     Header,
     Input,
     Label,
+    RadioButton,
+    RadioSet,
     RichLog,
     Rule,
     Select,
@@ -143,7 +145,7 @@ class SetupPane(Static):
 # ── Files list modal ─────────────────────────────────────────────────────────
 
 class FilesListScreen(ModalScreen):
-    """Modal showing all matched files with full paths."""
+    """Modal showing all matched files with extraction details on click."""
 
     DEFAULT_CSS = """
     FilesListScreen { align: center middle; }
@@ -152,17 +154,31 @@ class FilesListScreen(ModalScreen):
         border: thick $primary; background: $surface; padding: 1;
     }
     FilesListScreen DataTable { height: 1fr; }
+    FilesListScreen #extraction-panel {
+        height: auto;
+        max-height: 40%;
+        border: solid $primary;
+        padding: 0 1;
+        color: $success;
+        overflow-y: auto;
+    }
+    FilesListScreen #extraction-panel.muted { color: $text-muted; }
+    FilesListScreen #extraction-panel.error { color: $error; }
     FilesListScreen Horizontal { height: 3; align: right middle; }
     """
 
-    def __init__(self, files: list[str]) -> None:
+    def __init__(self, files: list[str], pattern: str = "", mode: str = "placeholder", fields: list[str] | None = None) -> None:
         super().__init__()
         self._files = files
+        self._pattern = pattern
+        self._mode = mode
+        self._fields = fields or []
 
     def compose(self) -> ComposeResult:
         with Vertical():
-            yield Label(f"[b]{len(self._files)} matched file(s)[/b] — full paths")
-            yield DataTable(id="files-table", zebra_stripes=True)
+            yield Label(f"[b]{len(self._files)} matched file(s)[/b] — click a file to see extracted fields")
+            yield DataTable(id="files-table", zebra_stripes=True, cursor_type="row")
+            yield Static("Select a file to see extracted fields.", id="extraction-panel", classes="muted")
             with Horizontal():
                 yield Button("Close", id="close-files", variant="primary")
 
@@ -171,6 +187,39 @@ class FilesListScreen(ModalScreen):
         table.add_column("File path")
         for f in self._files:
             table.add_row(f)
+
+    def on_data_table_row_highlighted(self, event: DataTable.RowHighlighted) -> None:
+        if event.row_key is None:
+            return
+        table = self.query_one("#files-table", DataTable)
+        filepath = str(table.get_row(event.row_key)[0])
+        self._show_extraction(filepath)
+
+    def _show_extraction(self, filepath: str) -> None:
+        from sovabids.parsers import parse_from_placeholder, parse_from_regex
+        panel = self.query_one("#extraction-panel", Static)
+        if not self._pattern:
+            panel.update("No pattern set.")
+            panel.remove_class("muted", "error")
+            panel.add_class("muted")
+            return
+        try:
+            if self._mode == "regex" and self._fields:
+                result = parse_from_regex(filepath, self._pattern, self._fields)
+            else:
+                result = parse_from_placeholder(filepath, self._pattern)
+            if result:
+                lines = "\n".join(f"  [cyan]{k}[/cyan] = [yellow]{v}[/yellow]" for k, v in _flatten_dict(result).items())
+                panel.update(f"[b]Extracted fields:[/b]\n{lines}")
+                panel.remove_class("muted", "error")
+            else:
+                panel.update("No fields extracted — pattern may not match this file.")
+                panel.remove_class("muted", "error")
[…]
