### 2026-05-05 aa73f4d
Enhance RulesPane to support file example mode for path pattern derivation and update UI elements accordingly
