+++ sovabids/tui.py
@@ -339,6 +339,12 @@ class RulesPane(Static):
     RulesPane #regex-fields-row.visible { display: block; }
     RulesPane #regex-examples { display: none; }
     RulesPane #regex-examples.visible { display: block; }
+    RulesPane #io-example-row { display: none; height: auto; }
+    RulesPane #io-examples { display: none; }
+    RulesPane #io-src-row { height: 3; }
+    RulesPane #io-src-row Input { width: 1fr; }
+    RulesPane #io-src-row Button { width: 18; margin-left: 1; }
+    RulesPane #pattern-section { height: auto; }
     RulesPane .inspect-row { height: 3; margin-bottom: 1; }
     RulesPane .inspect-row Button { margin-right: 1; }
     """
@@ -360,42 +366,70 @@ class RulesPane(Static):
         with RadioSet(id="pattern-mode"):
             yield RadioButton("Placeholder  (%field%)", id="mode-placeholder", value=True)
             yield RadioButton("Regex  (with named fields)", id="mode-regex")
+            yield RadioButton("File example  (source → BIDS path)", id="mode-example")
         yield Vertical(
             Label("Fields (comma-separated, same order as regex groups)"),
             Input(placeholder="entities.subject, entities.task", id="regex-fields-input"),
             id="regex-fields-row",
         )
+        yield Vertical(
+            Label("Source example  (one raw EEG file path)"),
+            Horizontal(
+                Input(placeholder="/data/sub-01_task-rest.vhdr", id="io-src-input"),
+                Button("Pick first match", id="io-pick-src", disabled=True),
+                id="io-src-row",
+            ),
+            Label("Target BIDS example  (corresponding BIDS output path)"),
+            Input(
+                placeholder="/bids/sub-01/eeg/sub-01_task-rest_eeg.vhdr",
+                id="io-tgt-input",
+            ),
+            id="io-example-row",
+        )
 
-        yield Label("[b]Path pattern[/b]", id="pattern-label")
-        yield Static(
-            "Use %entity% placeholders matched against the full file path.",
-            id="pattern-hint",
-            classes="hint",
+        yield Vertical(
+            Label("[b]Path pattern[/b]", id="pattern-label"),
+            Static(
+                "Use %entity% placeholders matched against the full file path.",
+                id="pattern-hint",
+                classes="hint",
+            ),
+            Static(
+                "[dim]Placeholder examples:[/dim]\n"
+                "  [cyan]%subject%_%task%.vhdr[/cyan]\n"
+                "    matches: /data/001_rest.vhdr  →  subject=001, task=rest\n"
+                "  [cyan]%subject%/%task%_%run%.bdf[/cyan]\n"
+                "    matches: /data/001/rest_01.bdf  →  subject=001, task=rest, run=01\n"
+                "  [cyan]raw/%subject%_%session%_%task%_%run%.eeg[/cyan]\n"
+                "    matches: /raw/001_ses01_rest_01.eeg\n"
+                "  [cyan]sub-%subject%/ses-%session%/eeg/sub-%subject%_ses-%session%_%task%_eeg.set[/cyan]\n"
+                "    already-partially-BIDSified paths",
+                id="placeholder-examples",
+                classes="pattern-examples",
+            ),
+            Static(
+                "[dim]Regex examples:[/dim]\n"
+                "  Pattern: [cyan]([^/]+)_([^/]+)\\.vhdr[/cyan]   Fields: [cyan]entities.subject, entities.task[/cyan]\n"
+                "    matches: /data/001_rest.vhdr  →  subject=001, task=rest\n"
[…]
