### 2026-05-05 df60de4
Add dependencies for TUI functionality and implement tests for TUI flow, directory picker, and files modal
