### 2026-05-05 7aa837b
refactor logging tests to check for streamhandler by name and add new test for stderr output with preexisting handler
