+++ tests/test_loggers.py
@@ -18,8 +18,8 @@ def test_setup_logging_adds_stderr_streamhandler(tmp_path):
     log_file = str(tmp_path / "test.log")
     setup_logging(log_file)
     root = logging.getLogger()
-    stream_handlers = [h for h in root.handlers if type(h) is logging.StreamHandler]
-    assert len(stream_handlers) >= 1
+    stream_handlers = [h for h in root.handlers if h.name == 'streamhandler']
+    assert len(stream_handlers) == 1
     assert stream_handlers[0].level == logging.WARNING
 
 
@@ -28,7 +28,7 @@ def test_setup_logging_stderr_handler_not_duplicated(tmp_path):
     setup_logging(log_file)
     setup_logging(log_file)
     root = logging.getLogger()
-    stream_handlers = [h for h in root.handlers if type(h) is logging.StreamHandler]
+    stream_handlers = [h for h in root.handlers if h.name == 'streamhandler']
     assert len(stream_handlers) == 1
 
 
@@ -62,6 +62,22 @@ def test_warnings_printed_to_stderr(tmp_path, capsys):
     setup_logging(log_file)
     logging.getLogger("sovabids.test").warning("console warning check")
     err = capsys.readouterr().err
+    assert "console warning check" in err
+
+
+def test_warnings_printed_to_stderr_with_preexisting_handler(tmp_path, capsys):
+    log_file = str(tmp_path / "sovabids.log")
+    root = logging.getLogger()
+    # Add a dummy stream handler that doesn't write to stderr
+    import io
+    dummy_stream = io.StringIO()
+    dummy_handler = logging.StreamHandler(dummy_stream)
+    root.addHandler(dummy_handler)
+    
+    setup_logging(log_file)
+    logging.getLogger("sovabids.test").warning("console warning check")
+    err = capsys.readouterr().err
+    # If setup_logging didn't add its own stderr handler, this will fail
     assert "console warning check" in err
 
 
