### 2026-05-05 fa014bb
refactor logging handlers to include 'sovabids_' prefix for clarity and update tests accordingly
