### 2026-05-05 85c5d82
refactor setup_logging to ensure single stderr handler and update tests for validation
