### 2026-05-05 4725ee9
Add TUI example for LEMON dataset usage
