### 2026-05-05 dacc308
Update README to include TUI installation and usage instructions
