### 2026-05-05 b9bb71b
Refactor README to enhance TUI and CLI usage instructions
