### 2026-05-05 06a8103
Update README to include additional examples for TUI and WEB GUI usage
