### 2026-05-05 88b3041
Add Quickstart guide and update README for new users
