### 2026-05-05 d3883ec
Clarify Mappings File description in documentation for better understanding of conversion parameters
