### 2026-05-05 058d07f
Add reference to new article by Dao et al. in research impact statement and bibliography
