### 2026-05-06 e82e0cb
Refine documentation in paper.md for clarity and accessibility of SOVABIDS features and interfaces
