### 2026-05-06 2207b60
Refine README links for improved clarity in TUI and WEB GUI examples
