### 2026-05-06 812fbf5
Enhance documentation and support for EEG/MEG formats in README, quickstart, and rules schema; update pyproject.toml for new dependencies and improve BIDS conversion logic in rules.py; add smoke tests for format conversions.
