+++ docs/source/quickstart.rst
@@ -1,7 +1,7 @@
 Quickstart
 ==========
 
-This page walks through a full EEG-to-BIDS conversion in one place.
+This page walks through a full EEG/MEG-to-BIDS conversion in one place.
 The four steps are always the same regardless of which interface you use:
 
 .. contents:: Steps
@@ -23,11 +23,16 @@ For CLI-only usage:
 
    pip install sovabids
 
+.. tip::
+   sovabids can read any EEG or MEG format supported by MNE's ``read_raw``. For native EDF or EEGLAB output,
+   install ``pip install "sovabids[formats]"``. For MEG data, set ``non-bids.output_format: FIF`` in your rules.
+   See the :ref:`Supported Formats <supported-formats>` section in the README for the full table.
+
 Step 2 — Prepare a rules file
 ------------------------------
 
 The **rules file** tells sovabids how to interpret your dataset as a whole —
-which EEG extension to look for, how to extract subject/session/task from file
+which file extension to look for, how to extract subject/session/task from file
 paths, power line frequency, channel corrections, and so on.
 
 See the full :doc:`rules_schema` for all options.
