### 2026-05-06 d6b9703
Add MEG-BIDS reference and update documentation for EEG/MEG support in SOVABIDS
