### 2026-05-06 f0ae065
Clarify format support details in paper.md, specifying read-only formats handled by MNE delegation and their testing status in SOVABIDS.
