### 2026-05-06 100a0dc
Refine documentation to clarify EEG/MEG support; update README, quickstart, and rules schema for experimental MEG features
