+++ docs/source/quickstart.rst
@@ -1,7 +1,7 @@
 Quickstart
 ==========
 
-This page walks through a full EEG/MEG-to-BIDS conversion in one place.
+This page walks through a full EEG-to-BIDS conversion in one place.
 The four steps are always the same regardless of which interface you use:
 
 .. contents:: Steps
@@ -24,9 +24,9 @@ For CLI-only usage:
    pip install sovabids
 
 .. tip::
-   sovabids can read any EEG or MEG format supported by MNE's ``read_raw``. For native EDF or EEGLAB output,
-   install ``pip install "sovabids[formats]"``. For MEG data, set ``non-bids.output_format: FIF`` in your rules.
-   See the :ref:`Supported Formats <supported-formats>` section in the README for the full table.
+   sovabids can read any EEG format supported by MNE's ``read_raw``. For native EDF or EEGLAB output,
+   install ``pip install "sovabids[formats]"``. MEG support is experimental — see the
+   :ref:`Supported Formats <supported-formats>` section in the README for details and caveats.
 
 Step 2 — Prepare a rules file
 ------------------------------
