### 2026-05-06 c17f62d
Enhance conversion error handling and logging; return success and failure details from convert_them function
