### 2026-05-06 6d0e161
Add tests for error handling during conversion; verify partial success and failure scenarios
