### 2026-05-06 4da10ce
Enhance tests for incremental conversion; verify that existing files are skipped during subsequent conversions
