### 2026-05-06 64c3138
Enhance documentation and examples for incremental conversion; clarify output handling and error reporting
