### 2026-05-06 8986dcf
Enhance conversion output to include skipped files; update documentation and examples accordingly
