### 2026-05-06 ef09d47
Update README and rules documentation to clarify automatic output format for MEG data; adjust rules logic to enforce FIF format
