### 2026-05-11 8e2f03e
Update link for Big Data Neuroscience Workshop poster
