+++ .github/workflows/python-tests.yml
@@ -60,7 +60,7 @@ jobs:
       if: success()
       uses: actions/upload-artifact@v4
       with:
-        name: uv-lock-${{ github.sha }}
+        name: uv-lock-${{ github.sha }}-py${{ matrix.python-version }}
         path: uv.lock
         retention-days: 90
 
