### 2026-07-14 yjmantilla/sovabids#89 @wmvanvliet
[JOSS] Rules file selector broken in TUI
In the TUI, on the setup screen, the file picker for the rules file picks the directory, not the actual file.
