### 2026-07-14 yjmantilla/sovabids#90 @wmvanvliet
[JOSS] textual not listed as dependency
The TUI depends on textual, but that is not listed in `pyproject.toml` as a required dependency.
