### 2026-07-14 yjmantilla/sovabids#91 @wmvanvliet
[JOSS] Contribution guide hard to find
The main docs make no mention of any contributor guidelines and after a bunch of searching I found the file tucked away in a hidden directory (`.github`). Why is it not more visible?
