### 2026-07-14 yjmantilla/sovabids#93 @wmvanvliet
[JOSS] Community guidelines missing
While there is a technical contribution guide, I can't find any community guidelines document as required by the JOSS review checklist:

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
