### 2026-07-14 55dc6eb
fix(docs): quote %-leading YAML pattern examples (#92)
A plain YAML scalar cannot begin with '%' (a reserved indicator character),
so the quickstart and rules_schema `pattern:` examples that start with
`%ignore%` / `%a%` failed to parse with a ScannerError. Quoting fixes it.

- Quote the two %-leading examples (quickstart, rules_schema operation example)
- Add a note explaining the quoting rule next to the %ignore% guidance
- Make the quickstart rules example a real fixture (examples/quickstart_rules.yml)
  rendered into the docs via `literalinclude`, so docs == tested file
- Add tests/test_do
[…]
