### 2026-07-24 yjmantilla/sovabids#94 @yjmantilla
Paths from the rules/mappings YAML do not expand $VARS or ~, silently creating a literal '$VAR' directory
## Summary

`General.IO.source` / `General.IO.target` are read straight out of the rules/mappings YAML and used as filesystem paths without expansion. If either contains an environment variable or `~`, sovabids does not fail — it creates a directory **literally named `$VAR`** (or `~`) relative to the current working directory and writes the BIDS output there.

`convert.py`:

```python
bids_path   = mappings['General']['IO']['target']
source_path = mappings['General']['IO']['source']
```

CLI usage is unaffected, because `sovabids rules <source_path> <bids_path> ...` gets its paths expanded by the shell before sovabids sees them. The exposure is specifically **paths that arrive via YAML**, which never touch a shell — and that is the documented way to drive a conversion.

## Reproduction

`mappings.yml`:

```yaml
General:
  IO:
    source: "$SOVASRC/srcdata"
    target: "$SOVAOUT/bidsout"
  dataset_description:
    Name: expansion-test
Individual: []
```

```bash
mkdir -p srcdata
SOVASRC=/tmp/sova_src SOVAOUT=/tmp/sova_out python -c \
  "from sovabids.convert import convert_them; convert_them('mappings.yml')"
```

**Expected:** output under `/tmp/sova_out/bidsout`.

**Actual:** `/tmp/sova_out` is never created; a directory named `$SOVAOUT` appears in the cwd. No exception.

```
$ ls -d */
$SOVAOUT/
srcdata/
```

`~` behaves the same way — `target: "~/sova_tilde_test/bidsout"` creates a directory literally named `~` in the cwd. That one is a small extra hazard to clean up, since `rm -rf ~` is not what you want to type.

The same applies to the lower-level entry point:

```python
from sovabids.bids import update_dataset_description
update_dataset_description({'Name': 'x'}, '$SOVATEST/bidsout')   # -> creates ./$SOVATEST/
```

Tested on `e562c1d`.

## Suggested fix

Normalise once, where YAML-supplied paths enter — `convert_them` and `apply_rules` — rather than at each `os.makedirs` call site:

```python
import os

def _resolve(p):
    return os.path.expandvars(os.path.expanduser(os.fspath(p))) if p is not None else p
```

`$SCRATCH`, `$SLURM_TMPDIR`, `$HOME` and `~` are the normal way to write portable paths in a config file, so this reads more as a missing capability than a niche edge case: a rules file that is meant to be shared between users cannot currently express "my scratch space" at all, it has to hard-code an absolute path per user.

If expansion is deliberately not wanted, the alternative that still removes the footgun is to **reject** paths matching `^[~$]` or containing `$` with a clear message, instead of silently materialising them.

Worth a regression test asserting no directory whose name starts with `$` or `~` is ever created.

## Context

Found this in [neurodags#28](https://github.com/yjmantilla/neurodags/issues/28) — same root cause, same silent failure mode — and checked whether sovabids shared it. It does. Flagging in case other tools in the stack take config-file paths the same way.
