### 2026-08-10 58cffa5
test(docs): run quickstart example end-to-end on a simulated dataset (#92)
The existing #92 guard proves the quickstart rules example *parses* (loads
through the real load_rules loader). This adds a test whose purpose is to
actually RUN the documented example: it loads the same examples/quickstart_rules.yml
fixture the docs render via literalinclude, simulates a tiny BrainVision dataset
laid out to match its "%ignore%/sub-%entities.subject%.vhdr" pattern, then runs
the quickstart's Steps 3-4 for real -- apply_rules (source -> mappings) and
convert_them (mappings -> BIDS) -- and asserts a valid
[…]
