### 2026-08-10 0d712ce
test(docs): drift-lock the rules-schema operation example via fixture (#92)
The quickstart example was already a single source of truth (docs literalinclude
the same examples/quickstart_rules.yml that a test loads). The second %-leading
site the reviewer named -- the operation example in rules_schema.rst -- was fixed
but not drift-protected: the test checked a hardcoded *copy* of the snippet, so a
future edit that unquoted the docs line would still ship green.

Make it a single source of truth too: extract the block to
examples/operation_example_rules.yml, render it via literalinclude, and loa
[…]
