### 2026-08-11 15f6aca
test(docs): exercise snippets through load_rules
