### 2026-08-13 317c75a
docs(contributing): surface CONTRIBUTING.md in rendered docs + README (#91)
The contribution guide lives at .github/CONTRIBUTING.md, which GitHub surfaces
in its Contribute/issue/PR UI, but it was invisible in the rendered docs and
unlinked from the README, so it was hard to find there.

- Add a docs/source/contributing.md stub that {include}s ../../.github/CONTRIBUTING.md
  (mirrors the existing gsoc_proposal.md precedent; no extra title so the guide's
  own "# Contributing to sovabids" H1 becomes the page title -> no duplicate-title
  warning), and add it to the index.rst toctree.
- Add a sh
[…]
