### 2026-08-14 8da2935
docs(community): add reporting/support/CoC guidelines + issue templates (#93)
The contribution guide covered only the technical dev workflow. JOSS wants
discoverable guidelines for all three of: contributing, reporting issues, and
seeking support. This adds the missing two plus a Code of Conduct, and surfaces
them the same way #91 surfaced the guide (docs stub + README), so they are
reachable from the rendered docs despite living in .github/.

- .github/CONTRIBUTING.md: add "Reporting Issues" (what to include: version +
  install method, OS/Python, interface, minimal rules snippet, full traceb
[…]
