+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,25 @@
+---
+name: Bug report
+about: Report a bug or unexpected behavior in sovabids
+title: ''
+labels: bug
+assignees: ''
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To reproduce**
+Please include as much of the following as you can:
+
+- sovabids version (`pip show sovabids`) and install method (pip / `.[dev]` / from source):
+- Operating system and Python version:
+- Interface used (Python API / CLI / TUI / JSON-RPC / web GUI):
+- A minimal rules file (or the relevant snippet) and the exact command or code you ran:
+- The full error message / traceback:
+
+**Expected behavior**
+What you expected to happen instead.
+
+**Additional context**
+Anything else that might help (sample data layout, screenshots, etc.).
