### 2026-08-14 3f2c6cc
docs(community): rename "Getting Help" -> "Seeking Support" (#93)
Match the JOSS checklist wording verbatim ("...Seek support") so the three
community-guideline sections map 1:1 to the checklist items a reviewer scans
for: Contribute / Reporting Issues / Seeking Support. Content unchanged; folds
in an impl-audit NICE-TO-HAVE (codex + grok).
