### 2026-08-18 yjmantilla/sovabids#91 @yjmantilla
You're right, it was basically invisible outside GitHub's own PR/issue prompts. The guide has been sitting at .github/CONTRIBUTING.md the whole time (that's why GitHub picks it up for the Contribute flow), but we'd never linked it from the rendered docs or the README, so there was no real way to find it there. My bad.

This is now fixed: it's a page in the docs nav (Contributing (https://sovabids.readthedocs.io/en/latest/contributing.html)), and the README has a Contributing section pointing at it. We kept the file under .github/ since that's a standard home for it, just made it reachable from the docs too. Thanks for flagging.
