### 2026-08-18 yjmantilla/sovabids#93 @yjmantilla
This is a real gap, yeah. What we had really only covered the dev side (fork, tests, PRs), nothing on reporting problems or getting help. I filled those in, and it's all reachable now from both the README and the rendered docs.▎

Contributing was already there. On top of that I added a section on reporting a bug (with what to include) that points at the issue tracker (https://github.com/yjmantilla/sovabids/issues), and one on getting help, where usage questions go to GitHub Discussions (https://github.com/yjmantilla/sovabids/discussions) or you can just email me... xD. I also added a Code of Conduct (https://github.com/yjmantilla/sovabids/blob/main/.github/CODE_OF_CONDUCT.md) and a bug-report template while I was at it. Thanks for pushing on this.
