### 2026-08-18 2717bd1
fix(tui): degrade sovatui gracefully when textual is missing (#90)
textual ships only in the optional [tui] extra, but sovatui was an
unconditional console script pointing at sovabids.tui:main, which imports
textual at module top level. So a plain `pip install sovabids` gave a raw
ModuleNotFoundError as soon as you ran sovatui.

- Add a thin entry module sovabids/sovatui.py: main() imports the real TUI
  lazily and, if textual is absent, exits with a clear "install
  sovabids[tui]" message (SystemExit, no traceback) instead of crashing.
  Repoint [project.scripts] sovatui to it. Guard is at th
[…]
