### 2026-08-18 9c200eb
test(tui): guard sovatui entry-point wiring + reframe paper TUI mention (#90 audit)
Impl-audit follow-up (all 3 panelists flagged the wiring gap):

- tests/test_sovatui_guard.py: add test_sovatui_entry_point_wired_to_wrapper,
  which reads [project.scripts] from pyproject.toml and asserts sovatui ->
  sovabids.sovatui:main. The runtime tests import the wrapper directly, so a
  revert of the script target back to sovabids.tui:main would leave them green
  while restoring the base-install crash; this closes that gap. Reads source
  pyproject (not importlib.metadata, which can reflect a stale ins
[…]
