### 2026-08-18 yjmantilla/sovabids#90 @yjmantilla
Good catch. textual is only in the optional [tui] extra, but sovatui gets installed as an entry point either way, so a plain pip install sovabids gave you a sovatui that just crashed with a ModuleNotFoundError.

The TUI is experimental and optional, so instead of making textual a hard dependency I kept it optional but made sovatui fail nicely, so if textual isn't installed it now prints an "install sovabids[tui]" message and exits without a traceback (with a regression test so it stays that way). I also went back through the docs and the paper and reframed the TUI as experimental rather than the default, with the CLI and Python API as the main interfaces.
