### 2026-08-18 4b17009
fix(tui): make the rules-file picker return a file, not a directory (#89)
The Setup tab's "Load existing rules file" browser reused the directory
picker, so it could only ever return a directory. Worse, MappingsPane then
silently ignored the non-file and built mappings from the Rules tab instead,
reporting success — so the bug was invisible at runtime (found via the TUI
bug-hunt panel, not just the reviewer's report).

- Add FilePickerScreen: a file browser where only a FileSelected sets the
  confirmable value; directory clicks navigate but never become the result,
  and OK is a no-op until o
[…]
