### 2026-08-18 c73bcd1
test(tui): make sovatui wiring test stdlib-only (fix CI on py3.9/3.10)
tomllib is 3.11+, so test_sovatui_entry_point_wired_to_wrapper failed under
CI's 3.9/3.10 with ModuleNotFoundError. Replace the tomllib parse with a
stdlib regex scan of [project.scripts] in pyproject.toml: same assertion
(sovatui -> sovabids.sovatui:main), works on every supported Python, and
still fails if the target is reverted to sovabids.tui:main.
