### 2026-08-18 56b6cdb
test(tui): harden #89 picker tests per impl audit
Impl-audit follow-up (all 3 panelists flagged test coverage of the core hazard):

- Post REAL DirectoryTree.FileSelected / DirectorySelected messages instead of a
  SimpleNamespace stub, so the tests also exercise Textual's event dispatch (a
  renamed handler / convention change would now be caught).
- Add test_tui_rules_picker_directory_click_cannot_override_file: choose a file,
  then navigate a directory, then OK -> the file survives. Also asserts
  FilePickerScreen has no on_directory_tree_directory_selected handler, locking
  in the struct
[…]
