### 2026-08-18 ab4a970
feat(tui): make the file/dir picker navigable (location bar + Up)
The pickers rooted a DirectoryTree at the seed path, so you could only expand
downward from there — if your file/folder wasn't under the starting directory
you couldn't reach it (reported while manually testing #89).

Factor both pickers onto a shared FSPickerScreen base that adds an editable
location bar (type/paste a directory, Enter to jump) and an Up button, both of
which re-root the tree (tree.path = _tree_root(path); tree.reload()).
DirPickerScreen / FilePickerScreen now only differ in their selection handler
and confirm r
[…]
