### 2026-08-18 c076f3c
feat(tui): double-click a folder to enter it in the pickers
Clicking a folder in a DirectoryTree only expands it inline; users expect a
double-click to navigate into it. Add double-click detection to the shared
FSPickerScreen: a second selection of the same folder within 0.5s re-roots the
tree into that folder. A single selection is delegated to a subclass hook —
DirPickerScreen marks it as the chosen directory, FilePickerScreen ignores it,
so a directory can still never become a file result (#89).

The #89 guarantee is now asserted behaviorally (a directory selection leaves
_selected empty in
[…]
