### 2026-08-18 34237d4
feat(tui): lock the Rules tab when a rules file is set in Setup
A rules file given in Setup wins over the Rules tab at Generate time
(#89), but nothing signalled that. Now setting the rules-file field
marks the Rules tab with a lock glyph, greys the pane, and shows a
banner saying it is ignored and how to undo it (clear the field).
Driven by Input.Changed, so the file picker triggers it too.
