### 2026-08-18 469af97
feat(tui): add a New folder button to the directory picker
The BIDS output directory usually doesn't exist yet, so the directory
picker now has a name box + New folder button that creates a folder in
the current location, steps into it, and pre-selects it. Dir-mode only;
the file picker (rules file) doesn't get it. Bad names/permission errors
surface in the picker status line instead of crashing.
