### 2026-08-18 9fcc62e
feat(tui): show example source paths in the Rules tab
Building a path pattern is guesswork without seeing real filenames, so
the Rules tab now lists a couple of actual source paths (relative, up to
two) for the chosen extension, refreshed when the tab opens or the
extension changes. Piggybacks on the existing file-count scan.
