### 2026-08-18 f71184e
feat(tui): move rules-file loader into the Rules tab; lock the builder; save feedback
Rules-related choices now live together. The 'Load existing rules file'
picker moves from Setup to the top of the Rules tab; loading a file greys
out the manual builder below it (in-tab lock, replacing the whole-tab lock)
and shows a note saying it's ignored. Setup is just source + output paths.

Ctrl+S 'Save rules' now gives feedback: a toast on success, and a warning
(instead of doing nothing) when the BIDS dir is unset or a rules file is
loaded. _generate reads the rules-file path from the Rules tab now.
