+++ sovabids/tui.py
@@ -275,18 +275,16 @@ class SetupPane(Static):
         with Horizontal(classes="path-row"):
             yield Input(placeholder="/path/to/bids_output", id="bids-input")
             yield Button("Browse…", id="browse-bids")
-        yield Rule()
-        yield Label("[b]Load existing rules file[/b] (optional — skips Rules tab)")
-        with Horizontal(classes="path-row"):
-            yield Input(placeholder="/path/to/rules.yml", id="rules-file-input")
-            yield Button("Browse…", id="browse-rules")
-        yield Static("Leave blank to build rules from scratch in the Rules tab.", classes="hint")
+        yield Static(
+            "Then define your conversion rules in the Rules tab "
+            "(or load an existing rules file there).",
+            classes="hint",
+        )
 
     def on_button_pressed(self, event: Button.Pressed) -> None:
         mapping = {
             "browse-source": ("source-input", True),
             "browse-bids": ("bids-input", True),
-            "browse-rules": ("rules-file-input", False),
         }
         if event.button.id in mapping:
             target_id, is_dir = mapping[event.button.id]
@@ -299,7 +297,6 @@ class SetupPane(Static):
         return {
             "source": self.query_one("#source-input", Input).value.strip(),
             "bids": self.query_one("#bids-input", Input).value.strip(),
-            "rules_file": self.query_one("#rules-file-input", Input).value.strip(),
         }
 
 
@@ -479,6 +476,19 @@ class RulesPane(Static):
         padding: 0 2;
         border-left: solid $primary;
     }
+    RulesPane .rulesfile-row { height: 3; }
+    RulesPane .rulesfile-row Input { width: 1fr; }
+    RulesPane .rulesfile-row Button { width: 14; margin-left: 1; }
+    RulesPane #rules-lock-note {
+        display: none;
+        height: auto;
+        color: $warning;
+        border: round $warning;
+        padding: 0 1;
+        margin-bottom: 1;
+    }
+    RulesPane #rules-lock-note.locked { display: block; }
+    RulesPane #rules-builder { height: auto; }
     RulesPane #ext-count { color: $text-muted; height: 1; margin-bottom: 1; }
     RulesPane #sample-paths {
         display: none;
@@ -527,6 +537,22 @@ class RulesPane(Static):
         self._matched_files: list[str] = []
 
     def compose(self) -> ComposeResult:
+        # Either load an existing rules file, OR build the rules below. Loading a
+        # file wins at Generate time (#89), so it locks the builder.
+        yield Label("[b]Load existing rules file[/b] (optional — locks the builder below)")
+        with Horizontal(classes="rulesfile-row"):
+            yield Input(placeholder="/path/to/rules.yml", id="rules-file-input")
+            yield Button("Browse…", id="browse-rules")
+        yield Static(
+            "🔒 A rules file is loaded — the builder below is ignored. "
+            "Clear the field to build rules here.",
+            id="rules-lock-note",
+        )
+        yield Rule()
+        with Vertical(id="rules-builder"):
+            yield from self._compose_builder()
+
+    def _compose_builder(self) -> ComposeResult:
         yield Label("[b]EEG file extension[/b]")
         ext_options = [(e, e) for e in SUPPORTED_EXTENSIONS]
         yield Select(ext_options, value=SUPPORTED_EXTENSIONS[0], id="ext-select")
@@ -632,8 +658,21 @@ class RulesPane(Static):
     # ── pattern preview ───────────────────────────────────────────────────────
 
     def on_input_changed(self, event: Input.Changed) -> None:
+        if event.input.id == "rules-file-input":
+            self._set_builder_locked(bool(event.value.strip()))
+            return
[…]
