+++ tests/test_tui.py
@@ -117,6 +117,8 @@ async def test_tui_rules_picker_returns_file(tmp_path):
     rules_yml.write_text("non-bids:\n  eeg_extension: .vhdr\n")
     app = SovabidsApp()
     async with app.run_test(size=(120, 40)) as pilot:
+        app.query_one("TabbedContent").active = "tab-rules"
+        await pilot.pause()
         await pilot.click("#browse-rules")
         await pilot.pause()
         # browse-rules opens the FILE picker, not the directory picker
@@ -144,6 +146,8 @@ async def test_tui_rules_picker_directory_click_cannot_override_file(tmp_path):
     sub2.mkdir()
     app = SovabidsApp()
     async with app.run_test(size=(120, 40)) as pilot:
+        app.query_one("TabbedContent").active = "tab-rules"
+        await pilot.pause()
         await pilot.click("#browse-rules")
         await pilot.pause()
         screen = app.screen
@@ -171,6 +175,8 @@ async def test_tui_picker_double_click_enters_folder(tmp_path):
     child.mkdir()
     app = SovabidsApp()
     async with app.run_test(size=(120, 40)) as pilot:
+        app.query_one("TabbedContent").active = "tab-rules"
+        await pilot.pause()
         app.query_one("#rules-file-input", Input).value = str(tmp_path)
         await pilot.pause()
         await pilot.click("#browse-rules")
@@ -237,6 +243,8 @@ async def test_tui_dir_picker_create_folder(tmp_path):
     # the FILE picker has no folder-creation controls (dir-only feature)
     app2 = SovabidsApp()
     async with app2.run_test(size=(120, 40)) as pilot:
+        app2.query_one("TabbedContent").active = "tab-rules"
+        await pilot.pause()
         await pilot.click("#browse-rules")
         await pilot.pause()
         assert isinstance(app2.screen, FilePickerScreen)
@@ -250,6 +258,8 @@ async def test_tui_rules_picker_ok_requires_file(tmp_path):
     from sovabids.tui import FilePickerScreen
     app = SovabidsApp()
     async with app.run_test(size=(120, 40)) as pilot:
+        app.query_one("TabbedContent").active = "tab-rules"
+        await pilot.pause()
         await pilot.click("#browse-rules")
         await pilot.pause()
         assert isinstance(app.screen, FilePickerScreen)
@@ -270,6 +280,8 @@ async def test_tui_rules_picker_reopen_from_file(tmp_path):
     rules_yml.write_text("a: 1\n")
     app = SovabidsApp()
     async with app.run_test(size=(120, 40)) as pilot:
+        app.query_one("TabbedContent").active = "tab-rules"
+        await pilot.pause()
         app.query_one("#rules-file-input", Input).value = str(rules_yml)
         await pilot.pause()
         await pilot.click("#browse-rules")
@@ -300,36 +312,63 @@ async def test_tui_generate_rejects_nonfile_rules(dummy_source, tmp_path):
 
 
 @pytest.mark.anyio
-async def test_tui_rules_file_locks_rules_tab(tmp_path):
-    """Setting a rules file in Setup locks the Rules tab (it loses at Generate time, #89);
-    clearing the field unlocks it again."""
-    from textual.widgets import TabbedContent
+async def test_tui_rules_file_locks_builder(tmp_path):
+    """Loading a rules file (in the Rules tab) greys out the manual builder below it;
+    it wins at Generate time (#89). Clearing the field unlocks the builder again."""
     rules_yml = tmp_path / "rules.yml"
     rules_yml.write_text("a: 1\n")
     app = SovabidsApp()
-    async with app.run_test(size=(120, 40)) as pilot:
-        tab = app.query_one(TabbedContent).get_tab("tab-rules")
-        banner = app.query_one("#rules-locked-banner", Static)
[…]
