### 2026-08-18 9ed1a9b
fix(tui): guard mappings save against writing to the CWD when BIDS dir is unset
Save Mappings YAML already reported status (Saved to … / Generate first),
but with the BIDS dir cleared, os.path.join("", ...) is relative and
silently wrote mappings.yml into the current working directory. Now it
warns instead, matching the Save Rules guard.
