### 2026-08-18 eddf71c
fix(tui): re-audit fixes — stale dir selection, folder-name escape, scan debounce
Independent adversarial re-audit of the post-#89 picker/rules UX work
(56b6cdb..HEAD) surfaced two confirmed correctness bugs and test gaps:

- DirPicker returned a stale selection: selecting a folder then navigating
  away (Up/location bar) still returned the earlier folder on OK. _reroot
  now clears the selection and OK label; double-click enters AND re-selects.
- New-folder box could escape its root via an absolute or ../ name
  (os.path.join). Now rejects names with / \ .. or absolute paths.
- ext-count/samp
[…]
