### 2026-08-18 yjmantilla/sovabids#99 @yjmantilla
TUI: preview workers race (unused `_preview_lock`) and leave stale action buttons after a failed rescan
# TUI: preview workers race (unused `_preview_lock`) and leave stale action buttons after a failed rescan

Two related defects in the Rules-tab live preview, both in `RulesPane`.

## 1. `_preview_lock` is declared but never acquired

`RulesPane` defines `_preview_lock: ClassVar = threading.Lock()`, but it is **never used** anywhere (grep:
exactly one occurrence, the declaration). The shared `self._matched_files` it is meant to guard is written
by `_update_show_files_btn` and read by the PSD / Channel-Names / Show-all / Pick-first actions with no
synchronization. The `@work(thread=True)` **preview** worker (`_preview_worker`) is not `exclusive`, and the
debounce timer (`_schedule_preview`) reduces but does not remove overlap, so results are
**last-completion-wins**: a slow scan for an older pattern can finish after a newer one and overwrite the
current file list.

> Note: the sibling **ext-count / sample-paths** worker was already made `@work(exclusive=True, group="ext-count")`
> and debounced on the #89 branch. This issue is the remaining half — the live **preview** worker
> (`_preview_worker`) plus the unused `_preview_lock`.

**Trigger:** type pattern A, immediately change to pattern B (or change source/extension) so A's scan
finishes after B's → the UI shows A's matches while the inputs describe B; the action buttons then operate
on the wrong file set.

## 2. Error paths leave stale action buttons enabled

On a `get_files` failure the preview worker sets an error message and `return`s **without** resetting
`_matched_files` / the action buttons. So after a previously successful scan, breaking the setup leaves
**Show all**, **Power Spectrum**, **Channel Names**, and **Pick first match** enabled and pointing at the
*previous* scan's files.

**Trigger:** get a good match (buttons enabled) → invalidate the source dir → rescan raises → preview shows
"Error" but the four buttons still act on the stale list.

## Evidence

`sovabids/tui.py`, `RulesPane`: `_preview_lock` declaration; `_preview_worker` / example-mode `get_files`
error returns; `_update_show_files_btn`; `_schedule_preview`.

## Suggested fix

- Make the preview worker `@work(thread=True, exclusive=True, group="preview")` (or cancel the in-flight
  worker before starting a new one), and either actually take `_preview_lock` around reads/writes of
  `_matched_files` or drop the unused lock.
- Call `self.app.call_from_thread(self._update_show_files_btn, [])` on **every** error / early-return path
  so a failed rescan clears the stale file list and disables the actions.

*Found by the TUI bug-hunt; see SYNTHESIS.md (finding 6).*
