### 2026-08-18 yjmantilla/sovabids#100 @yjmantilla
TUI: background workers query widgets and set `_mapping_data` off the UI thread
# TUI: background workers query widgets and set `_mapping_data` off the UI thread

## What happens

Several `@work(thread=True)` workers touch the widget tree or shared app state **on the worker thread**
instead of marshalling through `call_from_thread`:

- `MappingsPane._gen_worker` assigns `self.app._mapping_data = mapping_data` directly on the worker thread
  (only the table repaint after it is marshalled).
- `RulesPane._ext_count_worker`, `ConvertPane._convert_worker`, and the `ChannelNamesScreen` error path all
  evaluate `self.query_one(...)` on the worker thread to obtain a bound method, *before* `call_from_thread`
  runs.

Textual expects widget queries and state mutations to happen on the app (UI) thread. Querying/mutating from
a worker is a threading-contract violation: it can race a concurrent UI mutation, and it can target a
widget that has since been unmounted (e.g. a modal closed while its worker is still running).

## Why it matters

These are latent correctness/robustness bugs — intermittent and hard to reproduce, which is exactly why
they should be fixed structurally rather than left to chance. The off-thread `_mapping_data` write also
interacts with the stale-mappings issue (see the separate stale-mappings report).

## Evidence

`sovabids/tui.py`: `MappingsPane._gen_worker`; `RulesPane._ext_count_worker`; `ConvertPane._convert_worker`;
`ChannelNamesScreen` worker error path.

## Suggested fix

Resolve widgets on the UI thread and pass them in (or wrap the whole lookup-and-update in
`call_from_thread`), and assign `self.app._mapping_data` inside a main-thread callback rather than on the
worker. Do the work off-thread; touch the DOM/state only on the UI thread.

*Found by the TUI bug-hunt; see SYNTHESIS.md (finding 8).*
