### 2026-08-19 74d246f
test(tui): fix files-modal CI flake — tall viewport so #show-files is on-screen
The sample-paths block pushed #show-files to y=49, below the (120,40) test
viewport, so pilot.click hit OutOfBounds on CI (the earlier animated
scroll_visible raced and didn't settle before the click). Use a (120,100)
viewport so the button is on-screen without scrolling, and make the scroll
non-animated as belt-and-suspenders.
