### 2026-08-19 yjmantilla/sovabids#89 @yjmantilla
Yeah, this was broken. In the walkthrough video I just pasted the paths straight into the boxes, so I never actually used the Browse button and never hit the picker.

Fixed now: the rules-file browser is a file picker that returns the file you actually select, and there's a regression test for it now.

Also, I moved that "load existing rules file" box out of the Setup tab and into the Rules tab, where the rest of the rules stuff belongs.

Thanks for catching it.
