### 2026-08-19 38026a3
docs(tui): update the TUI tutorial for the new Rules-tab layout
The rules-file loader moved from the Setup tab to the top of the Rules tab
(loading a file now locks the builder instead of skipping the tab). Update
the tutorial accordingly, note the sample source paths shown in the Rules
tab and the new-folder option in the BIDS Browse dialog, and flag that the
walkthrough video shows the older layout.
