### 2026-08-19 d8e57ca
fix(tui): validate power-line frequency; drop dead _mappings reactive (#102)
An unparseable Power-Line-Frequency (e.g. "50hz") was silently dropped from
the built rules. The field now carries an optional-number validator (blank
still allowed), so a bad value is flagged instead of quietly omitted.

Also removed MappingsPane._mappings — a shared-mutable reactive([]) that was
never read or written (mappings live on app._mapping_data) — and its now
unused reactive import.
