### 2026-08-19 da00e07
fix(tui): make PSD preview cross-platform and stop leaking temp PNGs (#101)
_psd_worker hard-coded subprocess.Popen(["xdg-open", tmp]) (Linux only) and
wrote each PNG with NamedTemporaryFile(delete=False), never unlinked — so the
feature failed off Linux and every click leaked a temp file.

Add _open_path_in_viewer: dispatch to "open" (macOS) / os.startfile (Windows)
/ "xdg-open" (Linux), falling back to webbrowser file:// on any error. Reuse a
single per-pane temp file (psd.png) so clicks overwrite instead of accumulating,
and remove the temp dir in on_unmount. Path is resolved on the UI thre
[…]
