### 2026-08-19 0357cc4
fix(tui): stop preview-worker races and stale action buttons (#99)
Make _preview_worker @work(exclusive=True, group="preview") so a newer scan
cancels an older one instead of last-completion-wins overwriting the current
file list. On the get_files error paths (both example and normal mode), clear
_matched_files and disable the actions via _update_show_files_btn([]) so a
failed rescan can't leave Show-all/PSD/Channels/Pick-first pointing at the
previous scan. Drop the never-acquired _preview_lock and the now-unused
threading import.
