### 2026-08-19 6698df6
fix(tui): touch the DOM/shared state only on the UI thread in workers (#100)
Several @work(thread=True) workers evaluated self.query_one(...) or mutated
app state on the worker thread:

- _gen_worker assigned self.app._mapping_data on the worker; now set inside a
  call_from_thread callback (_apply_mappings) alongside the table repaint.
- _ext_count_worker resolved #ext-count off-thread to build a bound method;
  now marshals through _set_ext_count which queries on the UI thread.
- _convert_worker did log = self.query_one(#convert-log) on the worker; the
  RichLog is now resolved in _start_con
[…]
