### 2026-08-19 d2d0cdf
fix(tui): invalidate stale mappings; gate Save/Convert on a current plan (#98)
The generated plan on app._mapping_data bakes in the source/BIDS/rules at
generate time; nothing invalidated it when those inputs changed, so Convert
and Save Mappings could silently run against the old paths.

- Editing any Setup/Rules control invalidates the plan (clears _mapping_data
  and disables Save/Convert) via app-level input handlers scoped to Setup/Rules
  (modal picker inputs are excluded).
- _generate invalidates up front, so a generate that errors leaves no stale
  plan usable; the plan is (re)enabled 
[…]
