### 2026-08-19 4b1510d
fix(rules): surface the real cause when load_rules fails to parse YAML (#95)
load_rules caught everything with a bare except: and re-raised a generic
IOError("Couldnt read <path> file as a rule file."), discarding the real
error — so an unquoted %-leading pattern (a YAML ScannerError) showed up as
an unhelpful generic message. Keep the historical prefix (callers/tests match
it) but append the underlying error and chain it with 'from exc', and narrow
the bare except to 'except Exception'.
