### 2026-08-19 7b9c398
fix(rules): expand ~ and $VARS in source/BIDS/mappings paths (#94)
Paths taken from the rules/mappings YAML (or passed to the API) were used
literally, so a '~/data' or '$HOME/bids' silently became a directory named
'~' or '$HOME' — most visibly convert_them building the log dir from an
unexpanded target.

Add rules._expand_path (os.path.expanduser+expandvars; non-strings pass
through) and apply it at the path entry points: get_files (source),
apply_rules (source/bids/mapping_path), apply_rules_to_single_file
(file/bids), and convert_them (target/source + per-file IO.source/target).
Idempotent
[…]
