### 2026-08-19 c846ad1
fix(tui): block Generate/Save on a non-numeric power-line frequency (#102)
The validator only flagged an invalid PLF visually; get_rules still swallowed
the ValueError and Generate/Save proceeded, silently dropping PowerLineFrequency
(codex review). Add RulesPane.plf_error() and gate the builder path of _generate
and action_save_rules on it, surfacing an explicit message. Test now proves an
invalid value blocks Generate (not just that validate() returns invalid).
