### 2026-08-19 eef74c7
fix(tui): harden PSD opener + single-flight the preview (#101)
Per codex review: _open_path_in_viewer used fire-and-forget Popen, so an
xdg-open that exits non-zero (no display/handler) never triggered the fallback
while the UI reported 'opening'; the fallback URI f"file://{abspath}" is also
invalid on Windows/with spaces. Now run the opener (timeout) and check its exit
code, and fall back via Path(path).resolve().as_uri(). Also disable the Power
Spectrum button while its worker runs so a double-click can't race two writers
on the reused psd.png.
