### 2026-08-19 320bb33
fix(tui): drop stale preview results via a generation token; clear buttons on early returns (#99)
Per codex review: @work(exclusive=True) only cancels Textual's wrapper — the
thread keeps running, so a superseded _preview_worker could still land its
callbacks and overwrite the current file list/buttons. Add a _preview_gen
token bumped per scan; the worker reports results through _apply_preview(gen,
...) which drops them if a newer scan has superseded it. Also clear
_matched_files / disable the action buttons on _run_preview's early returns
(empty pattern, no source, missing regex fields / exam
[…]
