### 2026-08-19 a387092
fix(tui): drop a superseded generation via a token (no stale-plan resurrection) (#98)
Per codex review: invalidate-on-edit didn't help once a _gen_worker was already
in flight — its _apply_mappings callback re-stored the (stale) plan and re-enabled
Save/Convert, and two non-exclusive generations were last-wins. Add an app-level
_gen_token bumped on every invalidation (_set_mappings_valid(False), which fires on
edits and at _generate start). The worker captures the token and _apply_mappings /
_gen_error apply their result only while it's still current, so an edit or a newer
generation drops the
[…]
