### 2026-08-19 262fcb1
test(rules): cover per-file IO path expansion in convert_them (#94)
Per codex review: the existing convert_them test used an empty Individual, so
dropping the per-file IO.source/target expansion would still pass. Add a
non-empty mapping and assert apply_rules_to_single_file receives the EXPANDED
input and bids paths (no literal $VAR).
