### 2026-08-19 9b34534
fix(tui): reject non-finite/non-positive power-line frequency (#102 re-review)
float() accepts nan/inf/-5, which passed both the validator and plf_error. Add
_plf_problem() (positive + math.isfinite) shared by the validator and the gate.
