### 2026-08-19 f3e68e4
fix(tui): don't kill an attached viewer; robust PSD single-flight (#101 re-review)
Per re-review: subprocess.run(timeout=20) could wait on and then KILL an opener
that stays attached to the viewer. Use Popen + a short wait to detect an immediate
non-zero exit; a still-running opener is a successful hand-off (never killed).
_open_path_in_viewer now returns bool and the worker reports when nothing opened.
Replace the button-only single-flight with a _psd_in_flight flag so a preview
update (_update_show_files_btn) can't re-enable Power Spectrum mid-compute; resolve
the temp path before locking.
