### 2026-08-19 d0c6147
fix(tui): bump the preview generation on edit, not only at debounce (#99 re-review)
Per re-review: the token was bumped only when the debounced _run_preview fired,
so between an edit and the 600ms timer an in-flight worker could still apply
stale matches. _schedule_preview now bumps _preview_gen and clears the stale
matches/buttons immediately, then debounces only the replacement scan.
