### 2026-08-19 9807089
fix(tui): make Generate compute-only, not persist mappings.yml (#98 re-review)
Per re-review: apply_rules writes bids/code/sovabids/mappings.yml, so a superseded
or concurrent generation could overwrite/corrupt it even though its in-memory result
is dropped by the token. _gen_worker now hands apply_rules a private throwaway temp
mapping_path (removed in finally), so Generate touches only in-memory state; the
explicit Save Mappings action is the sole persistence point.
