### 2026-08-19 a9ab0bf
fix(tui): expand ~ and $VARS in Setup paths and rules-file (#94 re-review)
Per re-review: the TUI built paths (Save Rules/Mappings -> <bids>/code/...,
_generate's isfile check) from the raw Setup values, so a typed "~/bids" or
"$VAR/bids" created a literal directory. Expand in SetupPane.get_values (source,
bids) and RulesPane.get_rules_file so every TUI consumer sees the real path.
