### 2026-08-19 65325fb
fix(tui): reap a timed-out opener; guard PSD temp-path alloc (#101 re-review r3)
Per round-3 review: on the opener TimeoutExpired path the Popen was abandoned
without a later wait (possible zombie) — reap it in a daemon thread. And guard
_psd_png_path() (mkdtemp) in the show-psd handler so a temp-dir failure shows a
PSD error instead of crashing the app, leaving the single-flight lock false.
