### 2026-08-19 9ace969
fix(tui,rules): truly compute-only Generate; rescan preview on source change (#98 re-review r3)
Per round-3 review: temp mapping_path only stopped mappings.yml — apply_rules still
ran setup_logging (creating bids/code/sovabids + logs) and update_dataset_description
(rewriting an existing dataset_description.json) during mapping. Add persist=True to
apply_rules and apply_rules_to_single_file; persist=False skips the log setup, the
mappings write, and the dataset-description update (conversion is unchanged). _gen_worker
now calls apply_rules(persist=False) — touches nothing under bids, and drops
[…]
