### 2026-08-20 c28c80e
docs(paper): reduce overclaiming and tighten prose
Align claims with the code and remove overstatements flagged in an
editorial pass: drop the 'eliminates manual intervention', 'vast
majority', registry 'recognition', 'full provenance', MNE-BIDS
'scalability', RPC 'any application', and BIDScoin-novelty claims;
reword the read-only/BDF format note (BDF etc. are handled via MNE
delegation, just untested in CI) and state tested input vs default
output formats; align the summary and reproducibility wording with the
semi-automated, supervised-adjustment framing.

De-duplicate the YAML declarative 
[…]
