### 2026-08-20 #10146 @yjmantilla
Thanks @wmvanvliet and @rkurchin, and sorry for the delay. All five are fixed and closed now, all merged to main.

## Rules file selector broken in TUI (yjmantilla/sovabids#89)

Right, the "load existing rules file" browser was reusing the directory picker, so it just handed back a folder, not the file you clicked. In the walkthrough video I'd pasted the paths in by hand, so I never actually used the Browse button and never hit it. Fixed now: browse opens a file picker that returns the file you select (with a regression test), and I moved that box out of the Setup tab into the Rules tab where the rest of the rules stuff belongs.

## textual not listed as dependency (yjmantilla/sovabids#90)

Textual was only in the optional `[tui]` extra, but sovatui got installed either way. So on a plain install, running sovatui failed with a `ModuleNotFoundError`. I kept it optional but made sovatui fail with a "pip install sovabids[tui]" hint instead, and I reframed the TUI as an optional/experimental interface across the README, quickstart, and paper rather than the primary one.

## Contribution guide hard to find (yjmantilla/sovabids#91)

It was at .github/CONTRIBUTING.md but was never linked from the docs or the README, so it was basically invisible. Now it's a page in the docs nav, and the README has a Contributing section pointing at it.

## Quickstart example YML does not work (yjmantilla/sovabids#92)

Good catch, my bad! The pattern example started with an unquoted `%`, which YAML doesn't allow. Both %-leading examples are quoted now with a note explaining the rule, and they're covered by tests that load them, so a broken example shouldn't slip by silently again.

## Community guidelines missing (yjmantilla/sovabids#93)

I added a section on reporting a bug and one on getting help (Issues for bugs, Discussions for questions), plus a Code of Conduct and issue templates, all reachable from the README and the docs.

While I was in the TUI I also fixed a bunch of related stuff (path expansion yjmantilla/sovabids#94, worker threading yjmantilla/sovabids#99 and yjmantilla/sovabids#100, etc), all merged too.

Separately, I also did a pass over the paper text: tightened a few overstated claims and trimmed the prose. The full diff is in yjmantilla/sovabids@c28c80e if you want to look. Happy to adjust any of it. (cc @qian-chu, since this touches the paper text you already reviewed.)

The code changes are out as v0.4.6 on PyPI now.

Thanks again for the review.
