### 2026-08-30 #10146 @rkurchin
@yjmantilla we're into the home stretch! Please take a look at the author checklist above and send me the requisite information (latest version number and archive DOI) as well as confirmation of the other pieces as soon as you can!
