### 2026-08-31 979f703
docs(paper): update paper date to 31 August 2026
