+++ paper/paper.md
@@ -55,7 +55,7 @@ affiliations:
     index: 9
 
 
-date: 25 February 2026
+date: 31 August 2026
 bibliography: paper.bib
 ---
 
