### 2026-08-31 #10146 @yjmantilla
Thanks @rkurchin! On the checklist now. Before I cut the final release and archive it, I'm doing a quick round with the co-authors to confirm the author list, affiliations, and ORCIDs. I've given them about a week to reply. As soon as I hear back I'll make the release, archive it on Zenodo, and post the version number and DOI here.
